+++ Source/OxyPlot.Wpf/PlotView.cs
@@ -128,7 +128,7 @@ namespace OxyPlot.Wpf
         {
             var scale = base.UpdateDpi();
             this.RenderContext.DpiScale = scale;
-            this.RenderContext.VisualOffset = this.TransformToAncestor(Window.GetWindow(this)).Transform(default);
+            this.RenderContext.VisualOffset = this.TransformToAncestor(this.GetAncestorWindowFromVisualTree(this)).Transform(default);
             return scale;
         }
 
@@ -142,6 +142,22 @@ namespace OxyPlot.Wpf
             var exporter = new PngExporter() { Width = (int)this.ActualWidth, Height = (int)this.ActualHeight };
             var bitmap = exporter.ExportToBitmap(this.ActualModel);
             Clipboard.SetImage(bitmap);
+        }
+
+
+        /// <summary>
+        /// Returns a reference to the window object that hosts the dependency object in the visual tree.
+        /// </summary>
+        /// <returns> The host window from the visual tree.</returns>
+        private Window GetAncestorWindowFromVisualTree(DependencyObject startElement)
+        {
+            DependencyObject parent = startElement;
+            while (!(parent is Window))
+            {
+                if (parent == null) { break; }
+                parent = VisualTreeHelper.GetParent(parent);
+            }
+            return parent as Window ?? Window.GetWindow(this);
         }
     }
 }