+++ CHANGELOG.md
@@ -23,6 +23,7 @@ All notable changes to this project will be documented in this file.
 - Incorrect label placement on BarSeries with non-zero BaseValue (#1726)
 - LineAnnotation Text Placement on Reversed Axes (#1741)
 - Image opacity in WinForms and Core Drawing (#1766)
+- Fix specified Visual is not an ancestor of this Visual issue in WPF rendering (#1787) 
 
 ## [2.1.0-Preview1] - 2020-10-18